Refactored much in the tests
